[unit] fix unwrap
